Finish cakeday announcements
Lots of fiddling to get the formatting just right... landed on a dual approach where we use a single message for small # of users (common case) and split messages for lots of users.

Also fixed an issue about comparing local dates to UTC datetimes. Now everything is UTC, so the comparisons should work correctly.